+++ test/tst_dataview.jl
@@ -8,7 +8,6 @@
 
     @testset "constructor" begin
         @test_throws DimensionMismatch ObsView((rand(2,10),rand(9)))
-        @test_throws DimensionMismatch ObsView((rand(2,10),rand(9)))
         @test_throws DimensionMismatch ObsView((rand(2,10),rand(4,9,10),rand(9)))
         @test_throws MethodError ObsView(EmptyType())
         @test_throws MethodError ObsView(EmptyType(), ObsDim.Last())
@@ -178,6 +177,9 @@ end
     @test BatchView <: AbstractBatchIterator
     @test BatchView <: AbstractDataIterator
     @test batchview == BatchView
+    @test_throws MethodError oversample(BatchView(X))
+    @test_throws MethodError undersample(BatchView(X))
+    @test_throws MethodError stratifiedobs(BatchView(X))
 
     @testset "constructor" begin
         @test_throws DimensionMismatch BatchView((rand(2,10),rand(9)))